requirements.txt
tornado
line-bot-sdk
qrcode
tornado-redis
